chapter7/complete/sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swcpy"
version = "0.0.3"
authors = [
  { name="[Your Name]" },
]
description = "Example Software Development Kit (SDK) from Hands-On API Book for AI and Data Science"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
       'pytest>=8.1',
       'httpx>=0.27.0',
       'pydantic>=2.4.0',
       'backoff>=2.2.1',
       'pyarrow>=16.0',
       'python-dotenv',
]
